Move the neo4j to a store abstraction
