+++ src/test/java/no/ssb/dapla/blueprint/NotebookStoreTest.java
@@ -1,14 +1,14 @@
-package no.ssb.dapla.blueprint.parser;
+package no.ssb.dapla.blueprint;
 
 import no.ssb.dapla.blueprint.notebook.Notebook;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.neo4j.driver.Driver;
 import org.neo4j.driver.GraphDatabase;
 
 import java.util.List;
 
-class Neo4jOutputTest {
-
+class NotebookStoreTest {
     private static final List<String> DS_ONE = List.of(
             "/ds/one/one",
             "/ds/one/two",
@@ -30,11 +30,17 @@ class Neo4jOutputTest {
             "/ds/three/four"
     );
 
+    private NotebookStore store;
+
+    @BeforeEach
+    void setUp() {
+        Driver driver = GraphDatabase.driver("bolt://localhost:7687");
+        store = new NotebookStore(driver);
+
+    }
+
     @Test
     void testInsertNotebook() {
-        Driver driver = GraphDatabase.driver("bolt://localhost:7687");
-        Neo4jOutput output = new Neo4jOutput(driver);
-
         Notebook notebook = new Notebook();
 
         notebook.commitId = "commitId";
@@ -44,14 +50,11 @@ class Neo4jOutputTest {
         notebook.inputs = DS_ONE;
         notebook.outputs = DS_TWO;
 
-        output.output(notebook);
+        store.addNotebook(notebook);
     }
 
     @Test
     void testInsertNotebooksWithMatchingInputOutput() {
-        Driver driver = GraphDatabase.driver("bolt://localhost:7687");
-        Neo4jOutput output = new Neo4jOutput(driver);
-
         Notebook notebook = new Notebook();
 
         notebook.commitId = "commitId";
@@ -60,21 +63,19 @@ class Neo4jOutputTest {
         notebook.inputs = DS_ONE;
         notebook.outputs = DS_TWO;
 
-        output.output(notebook);
+        store.addNotebook(notebook);
 
         notebook.commitId = "commitId";
         notebook.path = "/some/other/path";
         notebook.fileName = "/some/other/path";
         notebook.inputs = DS_TWO;
         notebook.outputs = DS_THREE;
-        output.output(notebook);
+        store.addNotebook(notebook);
 
     }
 
     @Test
     void testInsertWithMatchingInputOutputDifferentCommits() {
-        Driver driver = GraphDatabase.driver("bolt://localhost:7687");
-        Neo4jOutput output = new Neo4jOutput(driver);
 
         Notebook notebook = new Notebook();
 
@@ -84,18 +85,14 @@ class Neo4jOutputTest {
         notebook.inputs = DS_ONE;
         notebook.outputs = DS_TWO;
 
-        output.output(notebook);
+        store.addNotebook(notebook);
 
         notebook.commitId = "commitId2";
         notebook.path = "/some/other/path";
         notebook.fileName = "/some/other/path";
         notebook.inputs = DS_TWO;
         notebook.outputs = DS_THREE;
-        output.output(notebook);
+        store.addNotebook(notebook);
 
     }
-
-
-
-
 }
